+++ CHANGELOG.md
@@ -20,7 +20,7 @@ That said, these are more guidelines rather than hard rules, though the project 
 
 ---
 
-## [0.13.0] - Unreleased
+## [0.13.0] - 2026-06-19
 
 ### Bug Fixes